+++ Questionnaire/Questionnaire.DAL/Entities/Data.cs
@@ -6,11 +6,9 @@ namespace Questionnaire.DAL.Entities
     [Table("Data")]
     public class Data
     {
-        public short Id { get; set; }
-        public int QuestionAnswerId { get; set; }
-       
-        public int FormId { get; set; }
-      
+        public int Id { get; set; }
+        public int QuestionAnswerId { get; set; }     
+        public int FormId { get; set; }     
         public string Comment { get; set; }
         
         public virtual QuestionAnswer QuestionAnswer { get; set; }
