+++ Questionnaire/Questionnaire.DAL/Entities/Question.cs
@@ -7,7 +7,7 @@ namespace Questionnaire.DAL.Entities
     public class Question
     {
         public int Id { get; set; }
-        public int QuestionTypeId { get; set; }
+        public short QuestionTypeId { get; set; }
         public string Name { get; set; }
 
         public virtual QuestionType QuestionType { get; set; }
